Additional Environmental Design
I have progressed with the continued design of the first scene. The scene will act as the long worn battlefield which the player will explore. Once this is done I can begin designing the 2nd scene which is mostly done in the other version of the Major Project. Once most of the environmental design is done I will begin importing some of the UI elements that I had created in the original project.