UI design
I have begun crreating my preliminary UI for the scene. The UI in this scene is meant to show off environmental storytelling by having UI test displayed over environmental objects like the dead bodies and the terrain objects. This is the method in which I will be telling my environmental storytelling instead of the 3D world text objects used in my original project.