Additional Assets and edits
I have added additonal assets to the scene and have begun changing up the inner castle level that i made  so that the design can be more linear and will allow for better player navigation. I have also seen that importing the asset from 1 project to another the textures are gone so what I will do is reapply the textures the next time i work on the project.