UI finishes and Menu work.
I've made the final adjustments to my UI test placements and how large they are meant to bo. The only thing I have left to do for the UI is make a count system so that the player is able to know how many they have found. I have also reworked certain elements for my Main Menu. I want to be able to show a small 3D environment so that the menu doesn't feel too empty.