Main Menu and Prefabs
I've made additional design edits to the main menu which includes taking aspects fo the first level. The way the menu is meant to work is that the player will sleect the level and then they will enter the scene almost as if they immedietly enter the game's level. I've also added additonal prefabs to help streamline and replace assets that I have made and that I can't improve with the time I have right now.