Additional UI work
I've made more progress on integrating the scroing system for the UI texts. I am also working on a pop-up panle for the pkayer to know what they are doing. I will have a Ui text asking players to look for a way into the castle and then if they need help on what to do or if they are confused there will be a tip icon which tells the players what optional UI texts they can find.