Scoring
I am implementing a system where the number of thoughts encountered will be counted so that the player will know how many they have found comapred to how many there are. There are still a few bugs to work out but I will get back to it on my next session.