Scene Switching and Scoring
I've set up the scene switching for the game and I have also turned off the Auto-Generate lighting function because it slows down the testing so I have decided to leave the generate lighting until after the environmental design side of the scenes have been set up. I have neglected to fix the scoring but I can fix that later.