Post Processing and UI design
I have begun adding post processing onto my scenes witha few mixed results. In the Inner Castle scene the scene and lighting is fine but the for the Outer Castle Level the lighting is a little too bright so that is something I will be fixing later on. I have begun redesigning and reintegrating the UI for the Inner Castle Level.