Collectible UI
I've been trying to reintegrate the collectible UI with a great success. The prefabs have been reintegrated and I have managed to set everything back up. The player when coming into conatct with the notes and candles will now pick up the collectibles and the buttons will come online. An issue I had with my first iteration is that the buttons are in the scene without having the notes picked up. This will be an issue that I need to fix in the future.