Collectible UI
I am still trying to fix a few issues with the collectible UI. At this point tyring to fix the issue has become a bit of a mess and has caused a few problems so my solution is to make a new mini project and remake the collectible system in that. I will then import that into this Project.