Collectible UI changes
I've made a few changes to the Ui and have undertaken a few tests that have revealed some anomolies. 1 major anomoly is that is full screen play mode the Panels for Inventory slots 2-5 open when clicked but don't close again unlike Panel 1 but is a smaller view the panels work perfectly.