Level Design Finishes
The Level design for the 2nd level is almost finished. There are a few blemishes left to clean up such as a little bit of positioning, certain area scalings and I need to layer my thoughts and collectbiles throughtout the level. I hasl also completley changed the layout of the castle.