Level Design finsihed and further UI design.
I have finished the Environmental/Level Design for the Inner Castle level. I am now setting up the Environmental and Collectible Storytelling elements via the UI system. Once this is done I will have 1 last scene to make which will be very small and be used to end the game and give resolution.