Level 1 and 2 are finished.
Levels 1 and 2 are relativly complete with all their UI text and collectbile systems placed and fixed, the environment completed and with scene transitioning. All that is left to do is fix the lighting and adjust the post-processing for both better visuals. Before moving on to the lighting and the Post-Processing I am going to make 1 more scene which consists of a road, a cave, a tree and some gold. After that is finished the Early Access version if the game will be finally finsihed.