Level 3 is Finished.
The level design for Level 3 is finished as well as the collectible and text UI. The lighting for the game has already been built. The level is intended to be a Prototype so in a further release version there will be a lot more detail and length to the Level.