Build Improvements
I was running throught the Prototype build I made and saw a number of issues such as problems with a few floor scalings and a few mistakes in the canvas's that have now been fixed. To the best of my abilities the Prototype is done and ready for another build and submission unless there are issues bought up in another meeting that need to be fixed before submission.