Final Submission Edits
I have made the final edits to the 2nd iteration of the Major Project and a build version has been made. I will go back to the project at a later date with massive plans to expand upon and greatly improve the project with systems such as puzzles, loading screens better stories.